+++ LiteCore/tests/CMakeLists.txt
@@ -51,13 +51,15 @@ file(GLOB FLEECE_FILES "../../vendor/fleece/Tests/*.json" "../../vendor/fleece/T
 file(COPY ${FLEECE_FILES} DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/vendor/fleece/Tests)
 add_executable(
     CppTests
+    LogEncoderTest.cc
+    ResultTest.cc
     c4BaseTest.cc
     c4DocumentTest_Internal.cc
+    AsyncTest.cc
     DataFileTest.cc
     DocumentKeysTest.cc
     FTSTest.cc
     LiteCoreTest.cc
-    LogEncoderTest.cc
     N1QLParserTest.cc
     PredictiveQueryTest.cc
     QueryParserTest.cc
